karsten:
randomfield.R: bandwidth now requires aws and uses geth.gauss to
determine bandwidth corresponding to the correlation.

lm.R: qlambda argument no longer used. new argument contrast which is a
vector and defualts to c(1). The vector will be filled with 0's up to
length ==  dim(z)[2] if neccessary or cut to this length. variance will
now contain the variance of this contrast of parameters and new value
cbeta contains the parameter contrast itself. value beta is unchanged.


git-svn-id: svn+ssh://localhost/Home/stat/WorkSpace/fmri-svn/trunk@32 090e7faf-2252-4674-a6b5-969773fc8662