author: Karsten
Added new dti module with 4 functions

1. create.designmatrix.dti()
   creates the matrix of the linear model for ln(s(b)/s0) from the
   gradient b vectors and the b-value.
2. calculate.lm.dti()
   estimate 6 tensor components from the linear model
   return, tensor elements, residualvarianz, xtx, and residuals (if
   needed)
3. determine.eigenvalue()
   calculate field of eigenvalues from tensor field.
4. anisotropy()
   calculate various anisotropy indices from the field of eigenvalues.

This module is still under construction. So, neither names nor location of
functions within the package are fixed. dti actually differs a lot from
fmri. We add it to the fmri package, since the aws procedures needed are
the same (cf. vaws3D). Note: a fucntion to determine the eigenvectors is
still missing.


git-svn-id: svn+ssh://localhost/Home/stat/WorkSpace/fmri-svn/trunk@43 090e7faf-2252-4674-a6b5-969773fc8662